requirements.txt
imageio>=2.31.2
matplotlib>=3.7.2
numpy>=1.25.2
Pillow>=10.0.0